requirements.txt
fastapi
uvicorn
SQLAlchemy
pymysql
alembic
pydantic
python-dotenv
pydantic-settings
cryptography
passlib
python-jose
pytest
httpx
pydantic[email]
requests
fastapi-mail
Jinja2
PyJWT
python-multipart